+++ django_browserid/__init__.py
@@ -5,7 +5,7 @@ This Source Code Form is subject to the terms of the Mozilla Public
 License, v. 2.0. If a copy of the MPL was not distributed with this
 file, You can obtain one at http://mozilla.org/MPL/2.0/.
 """
-__version__ = '0.7.1'
+__version__ = '0.8'
 
 from django_browserid.auth import BrowserIDBackend  # NOQA
 from django_browserid.base import BrowserIDException, get_audience, verify  # NOQA